backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
python-dotenv
httpx
stix2
taxii2-client
apscheduler
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==4.0.1
pycountry
